+++ lib/capybara.rb
@@ -84,6 +84,7 @@ module Capybara
     # [server = Symbol]  The name of the registered server to use when running the app under test (Default: :webrick)
     # [default_set_options = Hash]  The default options passed to Node::set (Default: {})
     # [test_id = Symbol/String/nil] Optional attribute to match locator aginst with builtin selectors along with id (Default: nil)
+    # [predicates_wait = Boolean]  Whether Capybaras predicate matchers use waiting behavior by default (Default: true)
     #
     # === DSL Options
     #
@@ -484,6 +485,7 @@ Capybara.configure do |config|
   config.reuse_server = true
   config.default_set_options = {}
   config.test_id = nil
+  config.predicates_wait = true
 end
 
 Capybara.register_driver :rack_test do |app|